interfazfinal/requirements.txt
Package                  Version
------------------------ ------------
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.7.0
h5py==3.10.0
pandas==2.2.1
numpy==1.26.2
PyTorch==2.4.1
python-dateutil == 2.8.2
python-dotenv }==  1.0.1
python-jwt    ==   2.0.1
torchvision==0.19.1
torchaudio==2.4.1
scikit-learn==1.5.2
scikit-image==0.24.0
seaborn==0.13.2
nltk==3.8.1
newspaper3k==0.2.8
oauth2client==3.0.0
matplotlib==3.8.4
redis==5.0.8
requests==2.32.3
sqlalchemy==2.0.31
python-dotenv==1.0.1
pytesseract==0.3.10  # Para OCR
opencv-python==4.9.0.80  # Para procesamiento de imágenes
tensorflow==2.17.0  # Si usas TensorFlow también
websocket-client==1.8.0
gcloud == 0.17.0
google-pasta == 0.2.0
googleapis-common-protos== 1.62.0